% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_the_best_abu_model.R
\name{find_the_best_abu_model}
\alias{find_the_best_abu_model}
\title{Best subset regression (BSR) on abundance data}
\usage{
find_the_best_abu_model(
  x,
  variables_lst = c("edge", "nbr", "understory", "kurtosis", "height90")
)
}
\arguments{
\item{x}{data frame that holds observation data and environmental metrics. observation data should be in a col called `obs`}

\item{variables_lst}{list of variables to include in BSR}
}
\value{
list of all models run
}
\description{
Best subset regression (BSR) on abundance data
}
